requirements.txt
asgiref
Django
djangorestframework
django-cors-headers
requests
sqlparse
tzdata
python-dotenv